+++ src/Fpl.Workers/Models/Mappers/StatHelper.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Fpl.Client.Models;
-using FplBot.Core.Models;
-using FplBot.Messaging.Contracts.Events.v1;
+﻿using FplBot.Messaging.Contracts.Events.v1;
 
 namespace FplBot.Core.Helpers
 {
@@ -10,25 +6,17 @@ namespace FplBot.Core.Helpers
     {
         public static StatType FromStatString(string identifier)
         {
-            switch (identifier)
+            return identifier switch
             {
-                case "goals_scored":
-                    return StatType.GoalsScored;
-                case "assists":
-                    return StatType.Assists;
-                case "own_goals":
-                    return StatType.OwnGoals;
-                case "yellow_cards":
-                    return StatType.YellowCards;
-                case "red_cards":
-                    return StatType.RedCards;
-                case "penalties_saved":
-                    return StatType.PenaltiesSaved;
-                case "penalties_missed":
-                    return StatType.PenaltiesMissed;
-                default:
-                    return StatType.Unknown;
-            }
+                "goals_scored" => StatType.GoalsScored,
+                "assists" => StatType.Assists,
+                "own_goals" => StatType.OwnGoals,
+                "yellow_cards" => StatType.YellowCards,
+                "red_cards" => StatType.RedCards,
+                "penalties_saved" => StatType.PenaltiesSaved,
+                "penalties_missed" => StatType.PenaltiesMissed,
+                _ => StatType.Unknown
+            };
         }
     }
 }
