+++ src/FplBot.Core/Helpers/Formatting/FixtureStats/IFormatEvents.cs
@@ -1,0 +1,10 @@
+namespace FplBot.Core.Helpers
+{
+    internal interface IFormatEvents : IFormat
+    {
+        string EventDescriptionSingular { get; }
+        string EventDescriptionPlural { get; }
+
+        string EventEmoji { get; }
+    }
+}
