+++ src/FplBot.Core/Helpers/Formatting/FixtureStats/NoOpFormatter.cs
