+++ src/FplBot.Core/Helpers/Formatting/FixtureStats/TauntType.cs
