+++ src/FplBot.Tests/Formatting/FixtureEventsFormattingTests.cs
@@ -4,7 +4,8 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using FplBot.Core.Abstractions;
 using FplBot.Core.Helpers;
-using FplBot.Core.Taunts;
+using FplBot.Core.Helpers.Formatting.FixtureStats;
+using FplBot.Core.Helpers.Formatting.Taunts;
 using FplBot.Data.Models;
 using FplBot.Messaging.Contracts.Events.v1;
 using FplBot.Tests.Helpers;
